+++ kubesae/utils.py
@@ -28,7 +28,7 @@ def get_backup_from_hosting(c, latest="daily", profile="caktus", backup_name=Non
     Params:
         c (invoke.Context): the running context
         latest (str, optional): Gets the latest backup from the specified temporal period. 
-            Defaults to "daily". Options are "daily", "weekly", "monthly", "yearly"
+            Defaults to "daily".
         profile (str, optional): The AWS profile to allow access to the s3 bucket. DEFAULT: "caktus"
         backup_name(str, optional): A specific backup filename.
         list(bool, optional): If set, will list the contents of the bucket for the projects folder and exit.
@@ -49,7 +49,6 @@ def get_backup_from_hosting(c, latest="daily", profile="caktus", backup_name=Non
         $ inv utils.get-db-backup --list --profile="client-aws"
             Will list all of the backup files using the a locally configured AWS_PROFILE named "client-aws"
     """
-    valid_periods = ['daily', 'weekly', 'monthly', 'yearly']
 
     if c.config.get("hosting_services_backup_bucket"):
         bucket = f"s3://{c.config.hosting_services_backup_bucket.strip('/')}"
@@ -70,10 +69,6 @@ def get_backup_from_hosting(c, latest="daily", profile="caktus", backup_name=Non
         c.run(f"aws s3 ls {bucket_folder}/ --profile {profile}")
         return
 
-    if latest not in valid_periods:
-        print(f"{latest} is not a valid backup interval. Valid intervals are {', '.join(valid_periods)}")
-        exit(1)
-
     if not backup_name:
         listing = c.run(
             f"aws s3 ls {bucket_folder}/ --profile {profile}",
@@ -86,13 +81,40 @@ def get_backup_from_hosting(c, latest="daily", profile="caktus", backup_name=Non
             for x in listing.split("\n")
             if re.search(f"^.*{latest}-.*", x)
         ]
+        if dates:
+            backup_name = f"{latest}-{c.config.hosting_services_backup_folder}-{dates[-1]}.pgdump"
 
-        backup_name = f"{latest}-{c.config.hosting_services_backup_folder}-{dates[-1]}.pgdump"
-    
+    if not backup_name:
+        print(f"No backup matching a latest of {latest} could be found.")
+        return
     c.run(
         f"aws s3 cp {bucket_folder}/{backup_name} ./{backup_name} --profile {profile}"
     )
 
+
+@invoke.task
+def list_backup_schedules(c, bucket_identifier='caktus-hosting-services-backups', profile='caktus'):
+    """
+    Lists the backup schedules found in a project's hosting bucket.
+    :param c:
+    :param bucket_identifier: The name of the bucket that holds the backups.
+    :param profile:
+    :return:
+    """
+
+    hosting_bucket = c.hosting_services_backup_folder if c.config.hosting_services_backup_folder else c.config.app
+    all_backups = c.run(
+        f"aws s3 ls s3://{bucket_identifier}/{hosting_bucket}/ --profile={profile}",
+        hide="out",
+    ).stdout.strip()
+
+    schedules = []
+    print(f"Backup schedules found at {hosting_bucket}\n")
+    for backup in all_backups.split("\r\n"):
+        name = backup.split(' ')[-1].split("-")[0]
+        if name not in schedules:
+            schedules.append(name)
+            print(f"{name}")
 
 @invoke.task
 def count_backups(c, bucket_identifier='caktus-hosting-services-backups', profile='caktus', extra_schedules=""):
@@ -126,3 +148,4 @@ def count_backups(c, bucket_identifier='caktus-hosting-services-backups', profil
 utils = invoke.Collection("utils")
 utils.add_task(get_backup_from_hosting)
 utils.add_task(count_backups)
+utils.add_task(list_backup_schedules)